+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 
 All notable changes to the Google Vision AI Bundle will be documented in this file.
 
-| **Bundle Version**  | **Pimcore Version** |
-|---------------------|---------------------|
-| 1.x.x               | Pimcore 10          |
+| **Bundle Version** | **Pimcore Version** |
+|--------------------|---------------------|
+| 1.x.x              | Pimcore 10          |
+| 2.x.x              | Pimcore 11          |
 
+## [2.0.0] - 2023-09-27
+- added Pimcore 11 compatibility
 ## [1.0.3] - 2023-03-09
 - update version in bundle class
 ## [1.0.2] - 2023-03-09